feat(circle-packing): move stories to the dedicated storybook workspace
